Fix issue with multiple SVG sprites
When provided with multiple SVG configs the script would always stop
at the first one. Moving the 'return' outside of the for loop fixed it.

Also adding a quick notify line which was helpful debugging which SVG
was being created. Decided to leave it there as it might be useful.